♻️[refactor] 탑뷰 리팩토링
